Added changes to trace plugin

+++ src/runtime_src/xdp/profile/database/static_info/aie_constructs.h
@@ -61,11 +61,11 @@ enum class module_type {
 
   struct tile_type
   { 
-    uint16_t row;
-    uint16_t col;
+    uint8_t row;
+    uint8_t col;
     uint16_t subtype;
-    uint16_t itr_mem_row;
-    uint16_t itr_mem_col;
+    uint8_t itr_mem_row;
+    uint8_t itr_mem_col;
     uint64_t itr_mem_addr;
     bool     is_trigger;
